separate code to components
